chore(gatsby): Convert utils/page-html to typescript
